refactor homescreen using enterscene/loading components
